pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
flask = "^3.0.0"
utils = "^1.0.1"
requests = "^2.31.0"
xds-protos = "^1.59.3"
torchvision = "^0.16.1"
mcr = "^0.1"
pillow = "^10.1.0"
sentence-transformers = "^2.2.2"
torch = "^2.1.1"
numpy = "^1.26.2"
opencv-python = "^4.8.1.78"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"